src/requirements.txt
bcrypt==3.1.7
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
colorama==0.4.3
colorlog==4.0.2
cryptography==2.8
idna==2.8
Jinja2==2.10.3
lxml==4.4.2
MarkupSafe==1.1.1
paramiko==2.7.1
prettytable==0.7.2
pycparser==2.19
PyNaCl==1.3.0
PyYAML==5.2
requests==2.22.0
six==1.13.0
urllib3==1.25.7
xmltodict==0.12.0
